move test dir out of src folder
