+++ packages/sdk/src/react/hooks/__tests__/useCurrencyBalance.test.tsx
@@ -1,7 +1,7 @@
 import { renderHook, waitFor } from '@test';
-import { TEST_ACCOUNTS } from '@test/const';
 import { zeroAddress } from 'viem';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { TEST_ACCOUNTS } from '../../../../test/const';
 import { useCurrencyBalance } from '../useCurrencyBalance';
 
 describe('useCurrencyBalance', () => {
